requirements.txt
flask
mysql-connector-python
os